requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
requests==2.31.0
python-whois==0.8.0
tldextract==5.1.1
Levenshtein==0.23.0
gunicorn==21.2.0